+++ packages/adp-tooling/src/prompts/index.ts
@@ -1,2 +1,3 @@
 export { getPrompts as getPromptsForChangeDataSource } from './change-data-source';
 export { getPrompts as getPromptsForNewModel } from './add-new-model';
+export { getPrompts as getPromptsForAddAnnotationsToOData } from './add-annotations-to-odata';