+++ packages/create/src/cli/add/index.ts
@@ -5,6 +5,7 @@ import { addAddCdsPluginUi5Command } from './cds-plugin-ui';
 import { addInboundNavigationConfigCommand } from './navigation-config';
 import { addCardsEditorConfigCommand } from './cards-editor';
 import { addNewModelCommand } from './new-model';
+import { addAnnotationsToOdataCommand } from './annotations-to-odata';
 import { addAddHtmlFilesCmd } from './html';
 
 /**
@@ -24,8 +25,10 @@ export function getAddCommands(): Command {
     addInboundNavigationConfigCommand(addCommands);
     // create-fiori add cards-editor
     addCardsEditorConfigCommand(addCommands);
-    // create-fiori add model
+    // create-fiori add new model
     addNewModelCommand(addCommands);
+    // create-fiori add annotations-to-odata
+    addAnnotationsToOdataCommand(addCommands);
     // create-fiori add html
     addAddHtmlFilesCmd(addCommands);
     return addCommands;